fix(neovim): loading error when lazyloading plugins
The issue was caused by loading configs of lazyloaded plugins

Therefore it was trying to load the config without the plugin being loaded first.

Solution:
- Remove lazyloading config for the affected plugins
- Add comment on add new plugins that require lazyloading and how to go about it

NOTE:
If the plugin needs to be lazyloaded, than make sure to provide the
config file as well. DON'T load the config in config.load-plugin-config file