fix(dotbot): installing homebrew with sudo
 # problem
When installing homebrew, it requires sudo access to be able to install
to their respective directories.
But, homebrew installation refuses to proceed if running the script AS
sudo.

 # solution
run the script with the variable INTERACTIVE=1 to force the script to as
for password, without the need to run as sudo